Refactor admin views (split into sub-modules)
