+++ cnxpublishing/views/admin/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013-2019, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+from __future__ import absolute_import
+
+from ..moderation import admin_moderations  # noqa: F401
+from ..api_keys import admin_api_keys  # noqa: F401
+from .content_status import *  # noqa: F401,F403
+from .index import *  # noqa: F401,F403
+from .post_publications import *  # noqa: F401,F403
+from .print_styles import *  # noqa: F401,F403
+from .site_messages import *  # noqa: F401,F403
