+++ cnxpublishing/views/admin/index.py
@@ -1,0 +1,38 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013-2019, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+from pyramid.view import view_config
+
+
+__all__ = (
+    'admin_index',
+)
+
+
+@view_config(route_name='admin-index', request_method='GET',
+             renderer="cnxpublishing.views:templates/index.html",
+             permission='preview', http_cache=60)
+def admin_index(request):  # pragma: no cover
+    return {
+        'navigation': [
+            {'name': 'Moderation List',
+             'uri': request.route_url('admin-moderation'),
+             },
+            {'name': 'API Keys',
+             'uri': request.route_url('admin-api-keys'),
+             },
+            {'name': 'Post Publication Logs',
+             'uri': request.route_url('admin-post-publications'),
+             },
+            {'name': 'Message Banners',
+             'uri': request.route_url('admin-add-site-messages'),
+             },
+            {'name': 'Content Status',
+             'uri': request.route_url('admin-content-status'),
+             },
+        ],
+    }
