+++ cnxpublishing/views/admin/post_publications.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013-2019, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+from __future__ import absolute_import
+
+from celery.result import AsyncResult
+from pyramid.view import view_config
+
+from ...db import db_connect
+
+
+__all__ = (
+    'admin_post_publications',
+)
+
+
+@view_config(route_name='admin-post-publications', request_method='GET',
+             renderer='cnxpublishing.views:templates/post-publications.html',
+             permission='administer', http_cache=0)
+def admin_post_publications(request):
+    states = []
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+SELECT ident_hash(m.uuid, m.major_version, m.minor_version),
+       m.name, bpsa.created, bpsa.result_id::text
+FROM document_baking_result_associations AS bpsa
+     INNER JOIN modules AS m USING (module_ident)
+ORDER BY bpsa.created DESC LIMIT 100""")
+            for row in cursor.fetchall():
+                message = ''
+                result_id = row[-1]
+                result = AsyncResult(id=result_id)
+                if result.failed():  # pragma: no cover
+                    message = result.traceback
+                states.append({
+                    'ident_hash': row[0],
+                    'title': row[1],
+                    'created': row[2],
+                    'state': result.state,
+                    'state_message': message,
+                })
+
+    return {'states': states}
