+++ cnxpublishing/views/admin/site_messages.py
@@ -1,0 +1,183 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013-2019, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+from __future__ import absolute_import
+from datetime import datetime, timedelta
+
+from pyramid import httpexceptions
+from pyramid.view import view_config
+
+from ...db import db_connect
+
+
+__all__ = (
+    'admin_add_site_message',
+    'admin_add_site_message_POST',
+    'admin_delete_site_message',
+    'admin_edit_site_message',
+    'admin_edit_site_message_POST',
+)
+
+
+@view_config(route_name='admin-add-site-messages', request_method='GET',
+             renderer='cnxpublishing.views:templates/site-messages.html',
+             permission='administer', http_cache=0)
+def admin_add_site_message(request):
+    banners = []
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+                SELECT id, service_state_id, starts, ends, priority, message
+                FROM service_state_messages ORDER BY starts DESC;""")
+            for row in cursor.fetchall():
+                banners.append({
+                    'id': row[0],
+                    'service_state_id': row[1],
+                    'starts': str(row[2]),
+                    'ends': str(row[3]),
+                    'priority': row[4],
+                    'message': row[5],
+                })
+    today = datetime.today()
+    tomorrow = today + timedelta(days=1)
+    return {'start_date': today.strftime("%Y-%m-%d"),
+            'start_time': today.strftime("%H:%M"),
+            'end_date': tomorrow.strftime("%Y-%m-%d"),
+            'end_time': tomorrow.strftime("%H:%M"),
+            'banners': banners}
+
+
+def parse_message_args(request):
+    args = {}
+    args['message'] = request.POST.get('message', 'Warning')
+    args['priority'] = request.POST.get('priority', 1)
+    args['type'] = request.POST.get('type', 1)
+
+    today = datetime.today()
+    tomorrow = today + timedelta(days=1)
+    start_date = datetime.strptime(
+        request.POST.get('start_date', today.strftime("%Y-%m-%d")),
+        '%Y-%m-%d').date()
+    start_time = datetime.strptime(
+        request.POST.get('start_time', today.strftime("%H:%M")),
+        '%H:%M').time()
+    end_date = datetime.strptime(
+        request.POST.get('end_date', tomorrow.strftime("%Y-%m-%d")),
+        '%Y-%m-%d').date()
+    end_time = datetime.strptime(
+        request.POST.get('end_time', tomorrow.strftime("%H:%M")),
+        '%H:%M').time()
+    start = datetime.combine(start_date, start_time)
+    end = datetime.combine(end_date, end_time)
+    args.update({'starts': start, 'ends': end})
+    return args
+
+
+@view_config(route_name='admin-add-site-messages-POST', request_method='POST',
+             renderer='templates/site-messages.html',
+             permission='administer', http_cache=0)
+def admin_add_site_message_POST(request):
+    # # If it was a post request to delete
+    # if 'delete' in request.POST.keys():
+    #     message_id = request.POST.get('delete', -1)
+    #     with psycopg2.connect(db_conn_str) as db_conn:
+    #         with db_conn.cursor() as cursor:
+    #             cursor.execute("""\
+    #                 DELETE FROM service_state_messages WHERE id=%s;
+    #                 """, vars=(message_id, ))
+    #     return_args = admin_add_site_message(request)
+    #     return_args['response'] = "Message id ({}) successfully removed".\
+    #                               format(message_id)
+    #     return return_args
+
+    # otherwise it was an post request to add an message banner
+    args = parse_message_args(request)
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+                INSERT INTO service_state_messages
+                    (service_state_id, starts, ends, priority, message)
+                VALUES (%(type)s, %(starts)s, %(ends)s,
+                        %(priority)s, %(message)s);
+                """, args)
+
+    return_args = admin_add_site_message(request)
+    return_args['response'] = "Message successfully added"
+    return return_args
+
+
+@view_config(route_name='admin-delete-site-messages', request_method='DELETE',
+             renderer='templates/site-messages.html',
+             permission='administer', http_cache=0)
+def admin_delete_site_message(request):
+    message_id = request.body.split("=")[1]
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+                DELETE FROM service_state_messages WHERE id=%s;
+                """, vars=(message_id, ))
+    return_args = admin_add_site_message(request)
+    return_args['response'] = "Message id ({}) successfully removed".\
+                              format(message_id)
+    return return_args
+
+
+@view_config(route_name='admin-edit-site-message', request_method='GET',
+             renderer='templates/site-message-edit.html',
+             permission='administer', http_cache=0)
+def admin_edit_site_message(request):
+    message_id = request.matchdict['id']
+    args = {'id': message_id}
+
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+                SELECT id, service_state_id, starts, ends, priority, message
+                FROM service_state_messages WHERE id=%s;
+                """, vars=(message_id, ))
+            results = cursor.fetchall()
+            if len(results) != 1:
+                raise httpexceptions.HTTPBadRequest(
+                    '{} is not a valid id'.format(message_id))
+
+            TYPE_MAP = {1: 'maintenance', 2: 'notice', None: 'maintenance'}
+            PRIORITY_MAP = {1: 'danger', 2: 'warning', 3: 'success',
+                            None: 'danger'}
+            args[TYPE_MAP[results[0][1]]] = 'selected'
+            args[PRIORITY_MAP[results[0][4]]] = 'selected'
+            args['message'] = results[0][5]
+
+            args['start_date'] = results[0][2].strftime("%Y-%m-%d")
+            args['start_time'] = results[0][2].strftime("%H:%M")
+            args['end_date'] = results[0][3].strftime("%Y-%m-%d")
+            args['end_time'] = results[0][3].strftime("%H:%M")
+    return args
+
+
+@view_config(route_name='admin-edit-site-message-POST', request_method='POST',
+             renderer='templates/site-message-edit.html',
+             permission='administer', http_cache=0)
+def admin_edit_site_message_POST(request):
+    message_id = request.matchdict['id']
+    args = parse_message_args(request)
+    args['id'] = message_id
+
+    with db_connect() as db_conn:
+        with db_conn.cursor() as cursor:
+            cursor.execute("""\
+                UPDATE service_state_messages
+                SET service_state_id=%(type)s,
+                    starts=%(starts)s,
+                    ends=%(ends)s,
+                    priority=%(priority)s,
+                    message=%(message)s
+                WHERE id=%(id)s;
+                """, args)
+
+    args = admin_edit_site_message(request)
+    args['response'] = "Message successfully Updated"
+    return args
